requirements.txt
telethon>=1.34,<2
python-dotenv>=1.0,<2

